backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
openai>=1.10.0
anthropic>=0.18.1
httpx>=0.26.0
pgvector>=0.2.4
asyncpg>=0.29.0
sqlalchemy>=2.0.25
aiofiles>=23.2.1
python-multipart>=0.0.6
google-generativeai>=0.3.2
